大幅なリファクタリング、特にファイル選別処理をクラス化した

+++ opus_conversion.py
@@ -7,11 +7,16 @@
 import os
 from os.path import expanduser
 import subprocess
-import ffmpeg
 from tqdm import tqdm
 import tkinter
 from tkinter import filedialog
-import concurrent.futures
+import datetime
+from multiprocessing import Process, Queue, Pool
+from queue import Empty
+import re
+from pathlib import Path, PurePath
+import signal
+
 
 # タグとカバーアート用のライブラリ群
 from io import BytesIO
@@ -22,76 +27,122 @@ from mutagen.flac import FLAC
 from mutagen.oggopus import OggOpus
 
 
+class WalkDirAndFiles:
+    """
+    ディレクトリーとファイルをリストアップたり、コピーしたりするクラス
+    """
+    def __init__(self, src:Path, dst:Path, target_file_type:list, dst_file_ext:list):
+        """
+        Args:
+        src:入力ディレクトリー(Pathオブジェクト)
+        dst:出力ディレクトリー(Pathオブジェクト)
+        target_file_type:入力のフィルターを掛ける拡張子
+        dst_file_ext:出力予定の拡張子
+        """
+        self.src_dir = src
+        self.dst_dir = dst
+        self.target_file_type = target_file_type
+        self.dst_file_ext = dst_file_ext
+        # ディレクトリーのリスト
+        self.src_dir_list = []
+        self.dst_dir_list = []
+        # 入力ファイルのリスト
+        self.src_file_list = []
+        # 出力ファイルのリスト　[入力ファイル, 出力ファイル１, 出力ファイル２, ....]
+        self.dst_file_list = []
+    
+    def walk_files(self, path:Path):
+        """
+        ディレクトリーを再帰的に検索、ファイルを見つけたら、file_listに追加する
+        また、入力するファイルのフィルターを掛ける
+        Args:
+        path:入力ディレクトリー(Pathオブジェクト)
+        file_list:pathの中に入っているファイルのリスト
+        """
+        for item in path.iterdir():
+            if item.is_dir():
+                self.walk_files(item)
+            elif item.is_file():
+                if item.suffix in self.target_file_type:
+                    self.src_file_list.append(item)
+    
+    def walk_dirs(self, path:Path):
+        """
+        ディレクトリーを再帰的に検索、フォルダーを見つけたら、dir_listに追加する
+        Args:
+        path:入力ディレクトリー(Pathオブジェクト)
+        dir_list:pathの中に入っているディレクトリーのリスト
+        """
+        for item in path.iterdir():
+            if item.is_dir():
+                self.src_dir_list.append(item)
+                self.walk_dirs(item)
+                
+    def make_list(self):
+        """
+        各種リストを作成
+        Args:
+        set_ext:出力するファイルの拡張子
+        filter_ext:フィルターする入力の拡張子
+        """
+        for dir in self.src_dir_list:
+            pure_dir = PurePath(dir)
+            sub_dir = pure_dir.relative_to(self.src_dir)
+            output_dir = PurePath(self.dst_dir)
+            self.dst_dir_list.append(output_dir / sub_dir)
+        
+        # 入力のファイルリストから出力のファイルリストを作成
+        for file in self.src_file_list:
+            # もし入力の拡張子と同じファイルがあれば
+            if file.suffix in self.target_file_type:
+                row_ext = []
+                row_ext.append(file)
+                for ext in self.dst_file_ext:
+                    file_name = PurePath(f"{file.stem}{ext}")
+                    pure_file = PurePath(file.parent)
+                    sub_file = pure_file.relative_to(self.src_dir)
+                    output_file = PurePath(self.dst_dir)
+                    row_ext.append(output_file / sub_file / file_name)
+                self.dst_file_list.append(row_ext)
+    
+    def make_dst_dirs(self):
+        """
+        出力先にディレクトリーを作成
+        Args:
+        """
+        for dir in self.dst_dir_list:
+            os.makedirs(dir, exist_ok=True)
+
+
+
 TARGET_FILE_TYPE = [".flac",".mp3"]
-OPUS_EXT = ".opus"
-WAV_EXT = ".wav"
-JPEG_EXT = ".jpg"
+DST_FILE_EXT = [".opus",".wav",".jpg"]
 
 # 入力するディレクトリーのパスと、その上のディレクトリーの名前を取得
 input_dir_pass = ""
-input_dir_root_pass = os.path.abspath(os.path.dirname(input_dir_pass))
 # 出力するディレクトリーのパスと、その上のディレクトリーの名前を取得
 output_dir_pass = ""
-output_dir_root_pass = os.path.abspath(os.path.dirname(output_dir_pass))
-
-# ファイルパスのリスト
-# (input_file_pass, output_file_pass, wav_file_pass, jpeg_file_pass)のタプルを入れる
-file_pass_list = []
 
 # Opusビットレート指定
 bitrate = "96"
 
-# マルチスレッド数指定
-max_thread = os.cpu_count() // 2
+# マルチプロセス数指定
+max_prosess_size = os.cpu_count() // 2
 
-def copy_directory(src, dst):
-    # この関数は、元のディレクトリー構造をコピー先に作り出します
-    # src:コピー元のディレクトリーのパス
-    # dst:コピー先のディレクトリーのパス
+def write_log(write_text_list):
+    # ログ用の初期化と作成
+    print("ログを書き出します")
+    now = datetime.datetime.now()
+    log_file_name = now.strftime("Opus_Conversion_log_%y-%m-%d_%H%M%S.txt")
+    os.makedirs("Log", exist_ok=True)
+    log_file_pass = os.path.join("Log", log_file_name)
     
-    os.makedirs(dst, exist_ok=True)
+    with open(log_file_pass, mode='w', encoding="utf-8") as f:
+        for text in write_text_list:
+            f.write(text + "\n")
     
-    # ディレクトリー構造をコピー
-    # もしすでにディレクトリーが存在したら、そのまま上書き
-    for root, dirs, files in os.walk(src):
-        for dir in dirs:
-            src_dir = os.path.join(root, dir)       # 入力のディレクトリー構造のパス
-            dst_dir = src_dir.replace(src, dst)     # 出力のディレクトリー構造のパス
-            os.makedirs(dst_dir, exist_ok=True)
-            
-        # ファイルが見つかった階層で、ターゲットファイルがあるかどうかチェック
-        if len(files) > 0:
-            target_file_check(src_dir, dst_dir)
-            
-def target_file_check(src_dir, dst_dir):    
-    # src_dirに存在する音楽ファイルのチェック
-    for root, dirs, files in os.walk(src_dir):
-        for input_file in files:
-            filename, ext = os.path.splitext(input_file)
-            if ext in TARGET_FILE_TYPE:
-                # もしターゲットの拡張子と一致したら
-                # 入力パスと出力パスのリストに追加
-                output_file = filename + str(OPUS_EXT)
-                wav_file = filename + str(WAV_EXT)
-                jpeg_file = filename + str(JPEG_EXT)
-                
-                abs_input_pass = os.path.join(input_dir_root_pass, os.path.join(src_dir, input_file))
-                abs_output_pass = os.path.join(output_dir_root_pass, os.path.join(dst_dir, output_file))
-                abs_wav_pass = os.path.join(output_dir_root_pass, os.path.join(dst_dir, wav_file))
-                abs_jpeg_pass = os.path.join(output_dir_root_pass, os.path.join(dst_dir, jpeg_file))
-                
-                file_pass_list.append((abs_input_pass, abs_output_pass, abs_wav_pass, abs_jpeg_pass))
+    print("ログの書き出しが完了しました。")
 
-def file_pass_check():
-    """
-    ファイルパスのチェック用関数
-    Args:
-    """
-    for list in file_pass_list:
-        print(f"input file pass:{list[0]}")
-        print(f"output file pass:{list[1]}")
-        print(f"wav file pass:{list[2]}")
-        print(f"jpeg file pass:{list[3]}")
         
 def export_coverart_img_and_tags(input_file_pass):
     """
@@ -102,6 +153,7 @@ def export_coverart_img_and_tags(input_file_pass):
     """
     filename, ext = os.path.splitext(input_file_pass)
     coverart_img = None
+    tags_dict = None
     
     # tags_dict = ファイルに添付されているタグの辞書
     
@@ -161,36 +213,46 @@ def convert_opus_func(file_pass):
     
     # インプットファイルを、メタデータとカバーアートを無視してPCM16bitに変換する
     # コンソール表示はOFFで実行
-    ffmpeg.input(input_file).output(wav_file, map_metadata='-1', acodec='pcm_s16le', vn=None).run(quiet=True)
+    ffmpeg_error = 0
     
+    command = ["ffmpeg", "-i", input_file, wav_file, "-map_metadata", "-1", "-acodec", "pcm_s16le", "-vn"]
+    try:
+        subprocess.run(command, stderr=subprocess.DEVNULL, check=True)
+    except subprocess.CalledProcessError as e:
+        ffmpeg_error = -1
+        print("ffmpeg error:" + str(e))
+            
     # Opusに変換する
-    subpro_opus(wav_file, output_file, bitrate, jpeg_file)
-    
-    # 生成した中間ファイルのWAVファイルを削除
-    if os.path.exists(wav_file):
-        os.remove(wav_file)
-    # 生成した中間ファイルのtemp.jpegを削除
-    if os.path.exists(jpeg_file):
-        os.remove(jpeg_file)
+    if ffmpeg_error == 0:
+        subpro_result = subpro_opus(wav_file, output_file, bitrate, jpeg_file)
         
-    insert_tags(output_file, tags)
+        # 生成した中間ファイルのWAVファイルを削除
+        if os.path.exists(wav_file):
+            os.remove(wav_file)
+        # 生成した中間ファイルのtemp.jpegを削除
+        if os.path.exists(jpeg_file):
+            os.remove(jpeg_file)
         
+        if re.search("success:", subpro_result):    
+            insert_tags(output_file, tags)
+            return subpro_result
+        elif re.search("error:", subpro_result):
+            return subpro_result
+        else:
+            print("error:conversion opus")
     
-def convert_opus_mt():
+def convert_opus_mt(file_path):
     """
     convert_opus_func()をマルチスレッドで回す
     Args:
         なし
     """
-    
-    tasks = file_pass_list
-    # 指定のスレッド数のスレッドプールを作成する
-    with concurrent.futures.ThreadPoolExecutor(max_workers=max_thread) as executor:
-        # 進捗管理のためtqdmでexecutor.mapの戻り値をラップ
-        # 進捗バーの表示更新のために、空のfor分を回す
-        for _ in tqdm(executor.map(convert_opus_func, tasks), total=len(tasks), desc="ファイルの進捗状況"):
-            pass
-        
+    result = []
+    with Pool(max_prosess_size) as pool:
+        result_imap = pool.imap(func=convert_opus_func, iterable=file_path)
+        for i in tqdm(result_imap, total=len(file_path), desc="ファイルの進捗："):
+            result.append(i)
+    return result
         
 def subpro_opus(input_wav, output_opus, bitrate="96", picture_pass = None):
     """
@@ -212,27 +274,42 @@ def subpro_opus(input_wav, output_opus, bitrate="96", picture_pass = None):
     # コマンドの実行
     try:
         subprocess.run(command, stderr=subprocess.DEVNULL, check=True)
+        success_text = "success:" + str(output_opus)
+        return success_text
     except subprocess.CalledProcessError as e:
-        print(f"エラーが発生しました: {e}")
+        error_text = "error:" + str(output_opus) + " : " +str(e)
+        return error_text
+
+def signal_handler(sig, frame):
+    exit("プログラムを強制終了します")
 
 def main():
+    
+    # はじめにCtrl+Cを押したときの動作を登録しておく
+    signal.signal(signal.SIGINT, signal_handler)
+    
     print("Opus変換プログラムへようこそ")
     input_check = input("入力するフォルダーを選択してください。続けるなら y を入力してください：")
     if input_check == "y":
         idir = os.path.abspath(expanduser("~"))
-        input_dir_pass = tkinter.filedialog.askdirectory(initialdir = idir)
+        input_dir_pass = Path(tkinter.filedialog.askdirectory(initialdir = idir))
         if input_dir_pass:
             output_check = input("出力するフォルダーを選択してください。続けるなら y を入力してください：")
             if output_check == "y":
-                output_dir_pass = tkinter.filedialog.askdirectory(initialdir = idir)
+                output_dir_pass = Path(tkinter.filedialog.askdirectory(initialdir = idir))
                 if output_dir_pass:
                     print("Opusへの変換を開始します")
-                    copy_directory(input_dir_pass, output_dir_pass)
-                    print(f"変換するファイルの合計：{len(file_pass_list)}")
+                    walk_dir_files = WalkDirAndFiles(input_dir_pass, output_dir_pass, TARGET_FILE_TYPE, DST_FILE_EXT)
+                    walk_dir_files.walk_dirs(input_dir_pass)
+                    walk_dir_files.walk_files(input_dir_pass)
+                    walk_dir_files.make_list()
+                    
+                    print(f"変換するファイルの合計：{len(walk_dir_files.src_file_list)}")
                     go_check = input("本当に変換しますか？実行するなら y を入力してください：")
                     if go_check == "y":
-                        # file_pass_check()
-                        convert_opus_mt()
+                        walk_dir_files.make_dst_dirs()
+                        log_result = convert_opus_mt(walk_dir_files.dst_file_list)
+                        write_log(log_result)
                         print("すべての処理が完了しました")
                     else:
                         print("変換を中止しました")
